MB-25183: Move decision of which seqno to purge upto out of seqlist
We want the 'list purge' to just do a generic purge of stale
elements upto a given seqno. The logic of not purging the last item
must be known only to the ephemeral vbucket as it is a vbucket
level constraint.

The vbucket passes a seqno upto which purge should happen and the
seqlist just purges the items till that point.

Change-Id: Ia4444689cfb79d182bd86eda52f16709e6e87888
Reviewed-on: http://review.couchbase.org/81877
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>