MB-24244: Ignore duplicate items during list range iteration
This commit modifies the list range iterator so that duplicate items
are not seen in the iteration range by a client like in-memory
buffered backfill.

As we iterate over the list, we check if a newer version is present
for a stale item within the iterator range (snapshot). If so, we
ignore reading the stale item and iterate over to the next item.

Example:
 A₁  B₂  C₃           Initial items
[A₁  B₂  C₃]          Range iteration 1-3
[A₁  B₂  C₃] B₄       Update B
 A₁  B₂  C₃  B₄       Range iteration ends
[A₁  B₂  C₃  B₄]      New range iteration 1-4, B₂ should be ignored
                      because B₄ is also in the snapshot
[A₁  B₂  C₃  B₄] B₅   Another update, but B₄ must still be sent as
                      B₅ is not in the range
 A₁  B₂  C₃  B₄  B₅   Range iteration ends

The range iteration client (1-4) sees [A₁ C₃  B₄], that is,  B₂ is
skipped.

Change-Id: I658f47d8eaf6b4e1a143809b9cd0f57f865ab266
Reviewed-on: http://review.couchbase.org/81932
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>