Use a common pattern to insert DCP packets to the output stream
This minimize the diff when we're trying to move over to cb::Pipe.
Almost all of the DCP packet producers start by formatting a
packet header which it wants to copy over into the destination
buffer. This patch call a method with the packet it want to
inject into the output stream, and let the method copy the
packet into the write buffer and add the pointer to the IO vector

Change-Id: I211cec563e800b7774cc039f751268bd3ec3543e
Reviewed-on: http://review.couchbase.org/82153
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>