MB-25179: Enable connection buffer recycling
Try to release the buffers assigned to each connection every
time a connection is put to sleep if the input or output
buffer is empty.

When the client connection is being scheduled, the framework
assigns an input buffer and an output buffer (the thread local
one is being used if one is available, otherwise a new
buffer is being allocated).

Before the connection goes to sleep all empty buffers would
be returned to the thread (or freed).

Change-Id: I7369e4743c1f04cc0b19b5e9eb4c9dc337fe41cb
Reviewed-on: http://review.couchbase.org/82154
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>