RocksDB: Introduce RocksDBKVStore
RocksDBKVStore is an experimental KVStore implementation backed by
RocksDB. It is very incomplete, and is not intended for general usage.

As such, it is hidden behind the `EP_USE_ROCKSDB` flag.

To use it, one can build with (for example)

make EXTRA_CMAKE_OPTIONS='-DEP_USE_ROCKSDB=ON'

As there is no UI involvement, buckets must be created as a normal
Couchbase bucket, and then have the backend changed with the following
diag/eval

 curl -X POST -u Administrator:asdasd http://localhost:9000/diag/eval \
 -d 'ns_bucket:update_bucket_props("bucket_name", [{extra_config_string,"backend=rocksdb"}]).'

Change-Id: I1353c9c86b4bfae86d98524cbcfe1a827ab5c703
Reviewed-on: http://review.couchbase.org/82164
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>