MB-25660: cbstats: More accurately calculate average for Histograms
The average calculation when printing histograms in cbstats is
incorrect - it under-accounts; sometimes significantly.

The problem is in the calculation of a representative value for each
datapoints range - the bin's count is multiplied by a representative
"value" for this data point, calculated as (end - start). This is
probably ok if the histogram bins grow exponentially - for example
with a power of 2:

    Start  End  Representative Value
        2    4  2
        4    8  4
        8   16  8

However, if the histogram bins grow slower, we end up with an
inaccurate representative value - for example a growth factor of 1.2:

    Start  End  Representative Value
        2    3  1
        3    4  1
        4    5  1

As a result the average calculated here will be incorrect - ~1 instead
of 3.5 (assuming 1 sample in each bin).

The solution is to just use the `start` value of the bin as the
representative value. I initially used the midpoint of start & end;
however that results in a massively inflated rep. value when we have
samples in the last histogram bucket; which has an end value of
max<TYPE> - for example:

   ro_1:fsReadCount (289135 total)
      ...
      8 - 10             : ( 28.20%)  77346 ########
      10 - 12            : ( 90.40%) 179835 #######################
      12 - 15            : ( 99.92%)  27536 ###
      ...
      22 - 26            : ( 99.99%)     10
      27173 - 4294967295 : (100.00%)     22
      Avg                : ( 163411)
   rw_1:fsReadSize (241768 total)

Clearly an Avg of 163411 is incorrect here!

An approximate set of affected histograms are:

    ‡ git grep Histo|grep Generator| grep -v ", 2)"
    engines/ep/src/hash_table.h:        : depthHisto(GrowingWidthGenerator<unsigned int>(1, 1, 1.3), 10),
    engines/ep/src/mutation_log.cc:    : paddingHisto(GrowingWidthGenerator<uint32_t>(0, 8, 1.5), 32),
    engines/ep/src/stats.h:        dirtyAgeHisto(GrowingWidthGenerator<hrtime_t>(0, ONE_SECOND, 1.4), 25),
    engines/ep/src/stats.h:        diskCommitHisto(GrowingWidthGenerator<hrtime_t>(0, ONE_SECOND, 1.4), 25),
    engines/ep/src/stats.h:        mlogCompactorHisto(GrowingWidthGenerator<hrtime_t>(0, ONE_SECOND, 1.4), 25),

Change-Id: Iecf9e75f77b13e68d9ab474864106251e75a1126
Reviewed-on: http://review.couchbase.org/82308
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>