MB-25630: Include read-only KVStore stats in 'cbstats kvtimings'
During investigation of slow background fetches on a customer
environment, I found that while we record filesystem timings for both
read-only and read-write KVStores, we only report the timings for
read-write.

To better assist in analysing read timings, we should also include the
read-only instance in the stats (accessed via the kvtimings cbstats
group).

Change-Id: Ie505e194cb1a91c80e8551cb198636ad68f56f1c
Reviewed-on: http://review.couchbase.org/82310
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>