Histogram stats: Don't omit empty bins with printing
When printing Histogram objects as stats (for display via cbstats),
don't omit histograms bins which contain zero samples.

While it initially seems sensible to ignore empty bins, when printing
histograms from such data it can be difficult to get a representative
view of the underlying distribution if empty bins are not printed in
the middle of a range.

For example, consider the following histogram which is from before
this patch:

 rw_0:snapshot (1296 total)
    2us - 4us        : (  0.08%)   1
    4us - 8us        : (  3.32%)  42 █▎
    8us - 16us       : ( 57.18%) 698 ██████████████████████▌
    16us - 32us      : ( 78.55%) 277 ████████▉
    32us - 64us      : ( 80.25%)  22 ▋
    1ms - 2ms        : ( 89.20%) 116 ███▊
    2ms - 4ms        : (100.00%) 140 ████▌
    Avg              : (  321us)

Note this is a bimodal distribution, with peaks at 8-16us and
2-4ms. However, from looking at the cbstats rendering of this it's not
obvious that there's actually a significant distance between the two
peak - approximately 3ms.

If we include empty (zero count) bins, we get a very different impression:
view:

 rw_0:snapshot (1296 total)
    2us - 4us        : (  0.08%)   1
    4us - 8us        : (  3.32%)  42 █▎
    8us - 16us       : ( 57.18%) 698 ██████████████████████▌
    16us - 32us      : ( 78.55%) 277 ████████▉
    32us - 64us      : ( 80.25%)  22 ▋
    64us - 128us     : ( 80.25%)   0
    128us - 256us    : ( 80.25%)   0
    256us - 512us    : ( 80.25%)   0
    512us - 1ms      : ( 80.25%)   0
    1ms - 2ms        : ( 89.20%) 116 ███▊
    2ms - 4ms        : (100.00%) 140 ████▌
    Avg              : (  321us)

Here the two discrete peaks are much clearer - we get an actual
quantatative representation of the data.

Implementation note: This increases the set of statistics seen by
ep_testsuite tests which verify that (a) all expected stats are output
and (b) no extra, unexpected stats are printed. Thus far we have
implemented these tests using an explicit whitelist of expected stat
names. However, with this change we now emit a much larger set of stat
names for each of the zero-count histogram buckets - e.g.

    rw_3:readSize_1048576,2097152
    rw_3:readSize_128,256
    rw_3:readSize_131072,262144

To avoid having to hardcode a large list of these kinds of stats -
whose keys will change if we adjust the histogram binning - this patch
adds support for specifying stat names using a regex pattern.

Given that these regexes cannot guard against extra stats being added
which happen to match the regex, this is only intended for
internal/auto-generated stat names which are *not* expected to be
exposed to end-users directly - we still want to maintain an explicit
set of user-visible stats.

Additionally, now that the full set of Histogram start/end ranges are
actually being emitted, ep_test_apis needs to be updated to handles
values up to uint64_t.

Change-Id: If2e9e420460fa185aff6c8e021800975d470b1d7
Reviewed-on: http://review.couchbase.org/82405
Reviewed-by: Oliver Downard <oliver.downard@couchbase.com>
Reviewed-by: Will Gardner <willg@rdner.io>
Tested-by: Build Bot <build@couchbase.com>