Fix race condition in default_engine:get_item_info
When a thread fetch the object it grabs the item mutex and
bumps the refcount of the item. That means that no one can
change the (user visible) properties on the underlying item.

Another thread may however grab the item mutex and toggle
the user invisible properties of the object (such as replacing
the value for a key by injecting a new item into the hashtable.

When we added support for locked items we started to look at
the iflag section in the item, which is one of the "internal
state variables" which could be modified while holding the
item mutex.

This patch change iflag to std::atomic as we don't want
to slow down get_item_info by locking the entire item
storage to check the internal flags.

This may potentially open up for a race, but:

1) If the item isn't linked anymore we don't need to mask
   the CAS anymore. (if the client tries to use that
   CAS it'll fail with an invalid cas)

2) In production the memcached buckets don't use the
   ZOMBIE state

Change-Id: I91818f6e43347691126337829555cb6495c235af
Reviewed-on: http://review.couchbase.org/82351
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>