Skip some of TestOperateOnStoredDocument tests under Thread Sanitizer
When running under Thread Sanitizer the named test takes roughly 20
seconds. Given that we run it 4 times (IPv4, IPv6, SSL IPv4 and
SSL IPv6) it is pretty slow. The test is all about checking that
memcached correctly checks the content of the document and only
allows increment/decrement if the value in the document consists
of whitespaced numbers.

Change-Id: I5407c73de89ceac92af72b73f7b295c0183997d3
Reviewed-on: http://review.couchbase.org/82355
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>