MB-25632: Allow slow op threshold to be customised
This patch removes the hardcoded threshold limits for various
commands, and place all of them in a JSON file with the following
syntax:

{
  "version": 1,
  "comment": "User supplied comment for the file (optional)"
  "default": {
    "slow": 500
  },
  "command-name": {
    "slow": 100
  },
}

The version field is mandatory and dictates the format of the fields
the parser should expect (to allow us to modify the content and still
be able to read the files in an upgrade scenario).

The comment is optional and ignored by the parser.

The field "default" is optional and contains the entry when there
isn't specified a value.

The rest of the file should contain entries per command. Each of
these entries contains a single field "slow" which contains the
threshold for that command (we might want to add other fields
later on).

If the value for slow is specified as a number, it is specified
in milliseconds. If the value is specified as a string it may
also contain one of the following specifiers:

    ns / nanoseconds
    us / microseconds
    ms / milliseconds
    s / seconds
    m / minutes
    h / hours

ex:

    "compact_db": {
       "slow": "30 m"
    }

One may also specify the values in memcached.json by providing
the same payload as described above in the key
"opcode_attributes_override".

Ex:

    ...
    "opcode_attributes_override": {
      "version":1,
      "default": {
        "slow": 200
      }
    }
    ...

The values in provided in memcached.json overrides any settings
in etc/couchbase/kv/opcode-attributes.json and
etc/couchbase/kv/opcode-attributes.d/<*.json>

Change-Id: I82b080705209632e88f128064a4f037f044209e5
Reviewed-on: http://review.couchbase.org/82450
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>