MB-25661: Statistics for background fetch read amplification
Add new statistics to ep-engine related to disk reads compared to the
data being read:

* ep_io_bg_fetch_read_count - Number of filesystem read()s issued
  while fetching documents from disk.

Using this statistic, we can derive a Read Amplification value for
background fetches:

* ep_bg_fetch_read_amplification - Ratio of filesystem read()s to
  number of documents fetched.

This can be plotted (e.g. with mortimer) to see the change over time,
but for convenience a statistic is provided to give the overall ratio
since last stat reset.

Change-Id: I7c77722ddccfb860ede6f159e04ffc7c814c1b34
Reviewed-on: http://review.couchbase.org/82569
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>