MB-24860 Cleanup thread-locals properly when ThreadLocal destructor runs
pthread_key_delete doesn't run destructors that were passed into
pthread_key_create. This causes issues for the thread which destructs a
ThreadLocalPosix object as it removes the thread-specific variable before
it's cleaned up its own thread-local which disables being able to clean
it up entirely.

Normally this isn't an issue as usually ThreadLocals get destructed by
a different thread (e.g. main) from the ones that actually use the
thread-specific variable (memcached worker threads / bucket creation
thread / ep-engine worker threads).

This is not the case in ep engine_testapp tests as the main thread both
creates an ep-engine instance and cleans it up which means the
destructor for the EPStats::localMemCounter is invoked before all the
threads that accessed it have exited.

The solution to this is to follow the instructions in the
pthread_key_delete man page and to manually invoke the destructor
(in the same way that the pthread_key_create man page describes it
as occurring).

This resolves a leak of 16 bytes that valgrind reports of a
`TLMemCounter` struct in ep engine_testapp tests.

Change-Id: Ia5626516b2440112687d91494de71fe47ef12bbe
Reviewed-on: http://review.couchbase.org/82535
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>