MB-25427 [1/4] Move GET_META to memcached - New functions in engines
GET_META command is currently implemented in engines (ep, default,
 ewouldblock) under `unknown_command()`. Memcached does not know
about GET_META, thus the MCBP request/response decoding/encoding
are delegated to the engines.
With this change we want to move all MCBP processing to memcached
and make the engines to peform only the bucket operations (i.e.,
retrieving the item metadata from memory or disk).
This first patch adds the new `get_meta()` functions into engines
and updates the engine testsuites to use the new API.
For now we do not switch the execution to the new implementation,
we will switch in a following patch.

Change-Id: Idda5e377d35a257287f39ef66e7915eb0ff22ef0
Reviewed-on: http://review.couchbase.org/82401
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>