Test malloc_usable_size is interposed
Adds a test to hopefully warn if we are using the system
malloc_usable_size (where available) but allocating memory with
jemalloc.

This should only be of importance if third-party code uses
malloc_usable_size - we should in kv_engine be using the existing
AllocHooks::get_allocation_size which maps to the correct call to the
underlying allocator, or returns a 0 if unavailable.

Change-Id: Ifd2beae94381c6ec86edf65e93469e67ccbdda7c
Reviewed-on: http://review.couchbase.org/82640
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>