MB-25786: Add $XTOC vattr to list XATTRs
This patch adds a new virtual xattr called $XTOC that contains an array
of xattr keys that exist on the document. E.g
{
    $XTOC: ["_sysXattr","userXattr"]
}

This vattr will only return the set of xattr keys that the user has the
rbac privileges for. For example if the user just has the
SystemXattrRead privilege then just the system xattr keys are listed in
the result. If the user has neither the XattrRead or the
SystemXattrRead privileges then EACCESS is returned.

Change-Id: I4ab204ac12749d5443701fdeba0ace4622a1cb27
Reviewed-on: http://review.couchbase.org/82642
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>