MB-25744: Update hashtable stats correctly during 'replaceByCopy'
In HashTable::unlocked_replaceByCopy(), we release the hash table
item and then insert its copy into the hash table. The stats that
are decremented in the release are not incremented during the copy.

This commit increments those stats which are decremented in
release. This is quite similar to how stats are incremented in
HashTable::unlocked_addNewStoredValue().

Change-Id: Ic310ce85d9b9b4ee7a6cacaf93b7cad7dbf6d139
Reviewed-on: http://review.couchbase.org/82603
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>