KVStore: move impl-only types to kvstore_priv.h
kvstore.h currently contains two classes of code - (1) the interface
for using KVStore subclasses, and (2) types and code used by
subclasses and to implement KVStore. This results in unnecessary code
being pulled in to files which just want to /use/ KVStore, and don't
need to implement it.

As such, create a new private header - kvstore_priv.h - to contain the
implementation details and move implementaiton code to this file.

Change-Id: Icfd721ffbb39005d7392f27806882d4b4ab97bc1
Reviewed-on: http://review.couchbase.org/82683
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>