Reduce coupling between vbucket.h and other classes
Move a number of VBucket inline methods to vbucket.cc; so clients of
vbucket don't need to know the details of the impementation (and hence
include those header files). Do a similar thing for KVBucket and
EPEngine.

Note: This is mostly a preparation patch; subsequent patches will
break the dependancy between vbucket.h and checkpoint.h.

Change-Id: I1080d5a51666c8bb82791c417504a5a830724f2f
Reviewed-on: http://review.couchbase.org/82693
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>