MB-25803: Fix race between streamRequest & VB deletion
There is a race condition which has been exposed by recent collections
changes. In DcpProducer::streamRequest we look up the VBucket object
for ther specified vbid, and if no such VBucket exists we return
ENGINE_NOT_MY_VBUCKET. We continue with all the various checks for the
request, and if successful, construct an ActiveStream object; passing
just the vbid (and *not* the VBucket ptr). In ActiveStream
constructor, we re-lookup the VBucket object for the given vbid.

However, if the VBucket object is deleted between the initial check in
the head of DcpProducer::streamRequest and ActiveStream constructor,
the VBucket for the given id has been deleted, and hence the second
lookup fails (and we end up with a null pointer).

Solution is to simply pass into ActiveStream to actual VBucket object;
and not re-acquire it. This ensures that the collections seperator can
safely be read.

Change-Id: I69825856117432aaf4cb46c6b641881747675c6b
Reviewed-on: http://review.couchbase.org/82727
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>