KVBucket: Use LockedVBucketPtr instead of direct use of vb_mutexes
Use the new RAII-style LockedVBucketPtr class for instances where we
need to acquire both a VBucketPtr and the associated vb_mutex.

In some instances we only want to attempt to lock the mutex, to
support this add a new variant of getLockedVBucket which will not
block (and return false) if the mutex could not be acquired.

Change-Id: I26a801a5639bc415cf19f30a9bf9f6253da0b3c2
Reviewed-on: http://review.couchbase.org/82736
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>