Warmup: Don't bgFetch an empty set of items
During Warmup via access log; if we end up with zero items in a batch
to fetch (because none of them are valid anymore) then don't attempt
to load zero items from disk.

As an additional check; in CouchKVStore::getMulti() return immediately
if we are passed an empty set of items to fetch.

While this is a sensible performance change, it was precipitated by a
divide-by-zero error at the end of CouchKVStore::getMulti() where we
use itms.size() as the divisor for a statistic calculation; assuming
that it would never be zero.

Change-Id: I2968e4c37eee07fe6ca2dca8e6e6caab7c78dca8
Reviewed-on: http://review.couchbase.org/82866
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>