MB-25820: Don't generate two responses for streamRequest
If a streamRequest is made to a dead vbucket we generated two results
(ENGINE_SUCCESS) followed by the correct result (NOT_MY_VBUCKET)

The reason was that we perform addFailoverLog, before checking if the
vbucket is dead.  addFailoverLog has the side-effect of invoking a
callback function, which is passed as a parameter to addFailoverLog.
The callback function invokes mcbp_response_handler and so generates
the response (ENGINE_SUCCESS).  We then continue the execution of the
streamRequest method and generate the second response
(ENGINE_NOT_MY_VBUCKET).

This patch separates out the invocation of the callback function from
addFailoverLog. (addFailoverLog is renamed getFailoverLog).
The callback invocation is then moved to after all of the error
checking cases.

A test is added to ensure that in the dead vbucket senario the callback
function is not invoked.

Change-Id: Ia6a711a36b9b964dcbd5c3cbd2baa5033389c09c
Reviewed-on: http://review.couchbase.org/82966
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>