MB-25894: Allow topkeys to be disabled through memcached.json
There is a mutex contention inside the topkeys implementation in
the server and the easiest solution is to allow people to
disable it if they want to squeeze the last bit of performance.

To disable topkeys support the following command may be used:

  curl -u Administrator:asdfasdf \
       -X POST \
       http://127.0.0.1:8091/pools/default/settings/memcached/node/self \
       --data topkeys_enabled=false

To enable, simply execute:

  curl -u Administrator:asdfasdf \
       -X POST \
       http://127.0.0.1:8091/pools/default/settings/memcached/node/self \
       --data topkeys_enabled=true

This patch opens up for ns_server doing "smarter" things later on.
It could be a button in the UI to enable it (could be located where
the table is), or ns_server could enable it once the user enters
the page etc.

Change-Id: I701aa0b73c231acd415b87e2580c2a1c102ec753
Reviewed-on: http://review.couchbase.org/83070
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>