Add new feature: UnorderedExecution
From the documentation:

`UnorderedExecution` - The client allows the server to reorder the
execution of commands (and send the responses in any order back to
the client). Note that when UnorderedExecution is selected, the
client cannot switch buckets (to make it deterministic which bucket
the operation is executed. The current proposal does not include any
barriers or other synchronization primitives.).

Change-Id: Ic3b93244d10ce09bef15c5c3596ca00af0dc9631
Reviewed-on: http://review.couchbase.org/83100
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>