MB-25881: Blob:getSize() to return the correct size
To ensure a blob size of 12 bytes, we pad by 3 bytes.
The getSize method did not take into account the padding and
therefore reported a size that was 3 bytes too big.

This patch changes the getSize method to account for the padding and
modifies an associated test that checks the value returned by getSize.

Change-Id: Ibd2e77b46ae221967e7a216715ab00a40a33e0d9
Reviewed-on: http://review.couchbase.org/83101
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>