Try to run more of the testsuites on top of ep engine
Up until now we've run the end to end tests (over the wire) purely
on default_engine. This have had the unfortunate side effect that
we haven't discovered the cases where ep-engine and default-engine
differ in return value for certain operations (we've had cases where
default_engine returns locked, and ep-engine returns tmpfail etc).

This patch adds an ENGINE parameter to specify the name of an engine
to test in _addition_ to default_engine. This patch should be modified
at some point by renaming ENGINE to ENGINES and contain a list of
all of the engines we want to run the test suite on top of.

Change-Id: Icd2a7b6fe184439d3fda17a227bc56d99bcb96db
Reviewed-on: http://review.couchbase.org/83124
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>