RocksDB: Fix memory leak in ~RocksDBKVStore
The destructor of RocksDBKVStore was not deleting the elements from
cachedVBStates, causing a memory leak on shutdown.

To solve this, change cachedVBStates from a non-owning vector to
std::vector<std::unique_ptr<vbucket_state>> so cleanup is handled
automatically.

As a side-effect, this highlighted that the KVStore & CouchKVStore
copy constructors were buggy, as they were not perfoming a deep-copy
of cachedVBStates (and hence would crash if one the objects was
deleted. Fortunatley nothing actually needed to copy CouchKVStore
objects, so just delete those constructors.

Change-Id: If8d3b659ac24f0f1820c08b7c5e3a66c0aa7980d
Reviewed-on: http://review.couchbase.org/83261
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>