MB-25991: Don't directly modify StoredValue during expiration
When performing the expiration of an item
(VBucket::deleteExpiredItem), is it not valid for us to directly
modify the StoredValue - as that skips updating the various counters
(numItems, numNonResident etc) which the HashTable needs to maintain.

Instead of ht.setValue (which should really be private), use
ht.unlocked_updateStoredValue() which performs the correct accounting.

Change-Id: Ia4e6112a9eda846ff8599ab985c8856f42d90ab2
Reviewed-on: http://review.couchbase.org/83357
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>