MB-26041: Add bucket name in the log message from EP::initialize
The bucket name isn't available before the provided
configuration is parsed. Given that we do log the configuration
if parsing fails, we can postphone logging the configuration
until we've actually parsed the configuration.

Given that the configuration is already parsed, the text
is changed from "Parsing config" to "Using configuration".

Change-Id: Ia0dd7755b675bd4d3ddfa56413f0930422317288
Reviewed-on: http://review.couchbase.org/83387
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>