MB-26037: Allow DCP rollback on vbuckets in pending state
We currently allow rollbacks on only replica vbuckets. This was
added recently (on 4.6.0 with http://review.couchbase.org/#/c/69725).

But during vbucket move, the new master is initially created in
pending state, items streamed over from old master and finally a
takeover stream is created. If there is a rollback during the
takeover, then we should still allow it on the vbucket which in the
pending state.

Change-Id: Ie57aed02d574be7b8a3852da5a948ef688676770
Reviewed-on: http://review.couchbase.org/83415
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>