MB-25342: Allow deletion of system_events
The next major collections patch changes how we manage the create and
delete markers in that a delete of a collection can be simply
represented as a delete of the create, exactly the same way a normal
user key is deleted.

This has one issue which this commit addresses in that Item can only
mark something as deleted by changing the operation, and for the
collection create/delete markers that is a problem as they are
represented by a special operation.

This commit removes the queue_op::del operation (and renames set to
mutation). This commit then adds a new boolean into Item in a way
that doesn't increase the sizeof(Item), now we can represent deleted
documents and deleted system events with ease and even backfill from
disk and correctly re-instate deleted system-events.

Change-Id: I5a743551e92f087cc96a06f40d08a3781c09314f
Reviewed-on: http://review.couchbase.org/83362
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>