MB-26126: HashTable::unlocked_restoreValue: fix deleted item counts
If a deleted document has its metadata restored back into the
HashTable (for example via GET_META), and then the whole value is
restored back (for example GET triggering a bgFetch), then the
numNonResident count is incorrectly decremented, and can underflow.

In more detail:

* GET_META("deleted_key") -> should result in HashTable state of:

  HashTable[0x104e0d818] with numInMemory:0 numDeleted:0 numNonResident:0 numTemp:1 values:
       SV @0x104ee0740 X.. ..N. temp: D  seq:-3 rev:2 key:"ns:0 key_1" exp:1505929057 vallen:0

* GET("deleted_key") -> should then result in HashTable with numNonResident==0:

  HashTable[0x104e0d818] with numInMemory:0 numDeleted:0 numNonResident:0 numTemp:1 values:
       SV @0x104ee0740 X.. WDNR temp:    seq:3 rev:3501 key:"ns:0 key_1" exp:4277009102 vallen:45 val:"x"

However we actually end up with a huge negative value:

  HashTable[0x104e0d818] with numInMemory:0 numDeleted:0 numNonResident:18446744073709551615 numTemp:1 values:
       SV @0x104ee0740 X.. WDNR temp:    seq:3 rev:3501 key:"ns:0 key_1" exp:4277009102 vallen:45 val:"x"

The issue is that deleted items are not correctly handled in
unlocked_restoreValue() - numItems is not incremented; nor is
numDeletedItems.

Fix this, and add additional tests in this area.

Change-Id: I41978112c80f091698b8368da9de8945ebfd412a
Reviewed-on: http://review.couchbase.org/83636
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Oliver Downard <olivermd@btinternet.com>
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>