MB-26197 [RocksDB] Add support for the Thread Pool options
RocksDB provides the `max_background_jobs`, `max_background_compactions`
and `max_background_flushes` options for tuning Multi-Thread Flush and
Compaction. With the first we just size the Thread Pool and let RocksDB
to decide how to allocate the available slots. The other two options are
deprecated but still in use in case the RocksDB default is not optimal.
For the last case we need to setup manually the HIGH Thread Pool for
Flush and the LOW Thread Pool for Compaction.

Change-Id: Ib98c079f18c554513727c4a774606452c7789eea
Reviewed-on: http://review.couchbase.org/83822
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>