MB-26255[2]: Don't use min compression ratio for compressing DCP document
During a backfill, we retrieve compressed documents from disk
anyways and there is no minimum compression ratio enforced. Hence,
don't enforce a compression ratio when compressing the document
before sending it over the wire. Ensure that the compressed length
is not bigger than the original length given that it doesn't help
reduce the network bandwidth.

Change-Id: I8c0bfb5645c0097a3c61742c29b7dd392c98d9c1
Reviewed-on: http://review.couchbase.org/83967
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>