Drop unused feature init_complete
The functionality was never enabled as it would require
bigger changes in ns_server to be useful in all scenarios.
As it stands today it is only usable in the scenario of
a cold startup of memcached. During other scenarios
(like rebalance) the node is already expected to be
"initialized" and it won't work in such situations.

Change-Id: I5eb098804a00668126c1b4cd3144e7bfaeea8ea3
Reviewed-on: http://review.couchbase.org/84017
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>