[1/n] MB-25920: Implement pause-resume capability in sequence list purge
EphTombstoneStaleItemDeleter task can have very long runtimes at both
on average and at tail end. The issue with that is not necessarily the
duration of the task itself; but it will block the other tasks in the
same class (NonIO) from running.

The biggest chunk of work performed by the task is iterating through
the entire sequence list and deleting stale items. This commit
adds capability to pause the tombstone purging while iterating through
the list and to resume purging from the paused point when purge is
called next.

And since purger tasks and snapshotted range read tasks are exclusive,
the snapshotted range reads should not have a problem.

(Periodic pause-resume at the task level will be implemented in the
 following commits)

Change-Id: I07e44b081348b568b2ec644f14966c3dc4bc095a
Reviewed-on: http://review.couchbase.org/83847
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>