Centralize HashTable count statistics
HashTable maintains a number of statistics about how many StoredValues
it contains which match some criteria - for example number of
non-resident, number with a given datatype, etc.

Ensuring the count of these statistics is accurate is currenly
error-prone - each of the different operations which may result in the
count changing needs to take care to update the relevent stat(s).

To make this more robust, refactor the counting of these statistics so
it is centrallised in two symmetric methods - statsPrologue() and
statsEpilogue(). statsPrologue should be called before making any
changes which may affect any of the stats (i.e. whenever a StoredValue
is changed), and statsEpilogue should be called afterwards.

By moving the all stats accounting to two methods (which are vitually
identical) we should minimise any accounting issues.

Change-Id: I3ffbb1c67eff078f21d62c3ed3a7fa958b8473b4
Reviewed-on: http://review.couchbase.org/84878
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>